+++ modules/sdnq/layers/linear/linear_int8.py
@@ -50,7 +50,17 @@ def int8_matmul(
 
 def quantized_linear_forward_int8_matmul(self, input: torch.FloatTensor) -> torch.FloatTensor:
     if torch.numel(input) / input.shape[-1] < 32:
-        return torch.nn.functional.linear(input, self.sdnq_dequantizer(self.weight, self.scale, self.zero_point, self.svd_up, self.svd_down, skip_quantized_matmul=True), self.bias)
+        dequantized_weight = self.sdnq_dequantizer(
+            self.weight,
+            self.scale,
+            self.zero_point,
+            self.svd_up,
+            self.svd_down,
+            skip_quantized_matmul=True,
+        )
+        if input.dtype != dequantized_weight.dtype:
+            input = input.to(dtype=dequantized_weight.dtype)
+        return torch.nn.functional.linear(input, dequantized_weight, self.bias)
     if self.sdnq_dequantizer.re_quantize_for_matmul:
         weight, scale = self.sdnq_dequantizer.re_quantize_matmul(self.weight, self.scale, self.zero_point, None, None)
         quantized_weight_shape = None